+++ spec/simple_crud_configuration_spec.rb
@@ -1,9 +1,15 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
+require 'simple_crud/authorization/pundit_adapter'
+require 'simple_crud/pagination/wor_paginate_adapter'
 
 class IncompleteAuthorizationAdapter
   include SimpleCrud::Authorization::Adapter
+end
+
+class IncompletePaginationAdapter
+  include SimpleCrud::Pagination::Adapter
 end
 
 describe SimpleCrud do
@@ -21,6 +27,16 @@ describe SimpleCrud do
 
       SimpleCrud::Config.authorization_adapter = original_adapter
     end
+
+    it 'lets a custom pagination adapter replace the wor-paginate default' do
+      original_adapter = SimpleCrud::Config.pagination_adapter
+      custom_adapter = instance_double(SimpleCrud::Pagination::WorPaginateAdapter)
+
+      described_class.configure { |config| config.pagination_adapter = custom_adapter }
+      expect(SimpleCrud::Config.pagination_adapter).to eq(custom_adapter)
+
+      SimpleCrud::Config.pagination_adapter = original_adapter
+    end
   end
 
   describe SimpleCrud::Authorization::Adapter do
@@ -32,6 +48,14 @@ describe SimpleCrud do
 
     it 'requires #policy_defined? to be implemented' do
       expect { adapter.policy_defined?(nil) }.to raise_error(NotImplementedError)
+    end
+  end
+
+  describe SimpleCrud::Pagination::Adapter do
+    subject(:adapter) { IncompletePaginationAdapter.new }
+
+    it 'requires #paginate to be implemented' do
+      expect { adapter.paginate(nil, nil, nil) }.to raise_error(NotImplementedError)
     end
   end
 end